+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 * **2019-01-26:**
     * Clone 5.10/unstable to 5.10/sc, update 5.10/sc VERSION to [5.10.10](https://community.ubnt.com/t5/UniFi-Beta-Blog/UniFi-Network-Controller-5-10-10-Stable-Candidate-has-been/ba-p/2649077)
+    * Update */root/usr/local/bin/entrypoint-functions.sh and docker-entrypoint.sh to 1.0.7
+        * Always chown on ${LOGDIR}/*.log and ${DATADIR}/system.properties when RUN_CHOWN=false. Fixes permission errors during clean OOB non-root startup with no existing data/dirs.
 ---
 * **2019-01-10:**
     * Clone 5.9/stable to 5.10/unstable, update 5.10/unstable VERSION to [5.10.5](https://community.ubnt.com/t5/UniFi-Beta-Blog/UniFi-Network-Controller-5-10-5-beta-has-been-released/ba-p/2627904)